+++ src/core/StorageService.js
@@ -415,7 +415,8 @@ exports.NewStorageService =
 							example: { foo: 'bar' },
 						} ),
 					],
-					view: 'storage/list',
+					// view: 'storage/list',
+					// view: 'Services/StorageService/List',
 				} );
 
 
@@ -442,7 +443,8 @@ exports.NewStorageService =
 							example: 'Read',
 						} ),
 					],
-					view: 'storage/item',
+					// view: 'storage/item',
+					// view: 'Services/StorageService/Item',
 				} );
 
 
@@ -462,7 +464,8 @@ exports.NewStorageService =
 							required: true,
 						} ),
 					],
-					view: 'storage/share',
+					// view: 'storage/share',
+					// view: 'Services/StorageService/Share',
 				} );